create R package skeleton
It has a completed DESCRIPTION file, NAMESPACE file generated by roxygen2
and a simple main.R file with a dummy function called swear_words. It is
the same function used here:
https://github.com/pdrhlik/southparkr/blob/master/R/analyses.R.

It just returns a vector of picked swear words.

Proper software engineering design will be needed so that the package can
actually be used in the future and that it can be easily usable and
extensible.