make tests local only #5
All the tests to ensure the right format of language files are in
data-raw/tests-local right now. These tests shouldn't be run on an
installed package. They wouldn't pass because the raw files aren't part of
the build. Only the resulting data frame is.

For now, I added a temporary empty test so that R CMD check can pass.