Close #21 - rename language files
When adding a new language, it is no longer necessary to edit README.Rmd.
We don't need to edit the language data frame any more because the
language name is present in the file name.

Example:
en_English
fr-CA_French (Canada)

The only issue that came up is that running Ctrl+Shift+K in Rstudio
(default Knit) doesn't work because of an unexported function load_langs.
It works when using build_sweary() so it's okay. This function cannot be
exported because it loads raw language files that are not present in an
installed package.